requirements.txt
streamlit
io
PLT
google-generativeai